split app settings in sperate module
