+++ starsky/starsky/clientapp/src/components/organisms/preference-app-settings-desktop/preference-app-settings-desktop.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const PreferencesAppSettingsDesktop: React.FunctionComponent = () => {
+  return (
+    <>
+      <p>Desktop</p>
+      Only used when using Starsky as desktop
+    </>
+  );
+};
+
+export default PreferencesAppSettingsDesktop;
