+++ starsky/starsky/clientapp/src/components/organisms/preferences-app-settings-storage-folder/preferences-app-settings-storage-folder.stories.tsx
@@ -1,0 +1,11 @@
+import PreferencesAppSettingsStorageFolder from "./preferences-app-settings-storage-folder";
+
+export default {
+  title: "components/organisms/preferences-app-settings-storage-folder"
+};
+
+export const Default = () => {
+  return <PreferencesAppSettingsStorageFolder />;
+};
+
+Default.storyName = "default";
