+++ starsky/starsky/clientapp/src/interfaces/IAppSettings.ts
@@ -1,5 +1,10 @@
+import { IAppSettingsDefaultEditorApplication } from "./IAppSettingsDefaultEditorApplication";
+import { RawJpegMode } from "./ICollectionsOpenType";
+
 export interface IAppSettings {
   verbose: boolean;
   storageFolder: string;
   storageFolderAllowEdit: boolean;
+  defaultDesktopEditor: IAppSettingsDefaultEditorApplication[];
+  desktopCollectionsOpen: RawJpegMode;
 }
