+++ starsky/starsky/clientapp/src/interfaces/IAppSettingsDefaultEditorApplication.ts
@@ -1,0 +1,6 @@
+import { ImageFormat } from "./IFileIndexItem";
+
+export interface IAppSettingsDefaultEditorApplication {
+  applicationPath: string;
+  imageFormats: ImageFormat[];
+}
