+++ starsky/starsky/clientapp/src/interfaces/ICollectionsOpenType.ts
@@ -1,0 +1,5 @@
+export enum RawJpegMode {
+  Default = 0,
+  Jpeg = 1,
+  Raw = 2
+}
